+++ cpp-ch/local-engine/Parser/SerializedPlanParser.h
@@ -187,13 +187,10 @@ static const std::map<std::string, std::string> SCALAR_FUNCTIONS
        // array functions
        {"array", "array"},
        {"size", "length"},
-       {"get_array_item", "arrayElement"},
-       {"element_at", "arrayElement"},
        {"range", "range"}, /// dummy mapping
 
        // map functions
        {"map", "map"},
-       {"get_map_value", "arrayElement"},
        {"map_keys", "mapKeys"},
        {"map_values", "mapValues"},
        {"map_from_arrays", "mapFromArrays"},
